requirements.txt
# Point Cloud Processing
laspy>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
open3d>=0.18.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning & Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Augmentation
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
h5py>=3.8.0

# Optional: For advanced point cloud processing
# pyntcloud>=0.3.0
# pclpy  # Python bindings for PCL (Point Cloud Library)
